fixed null-pointer constants and SkipToNewLine()

+++ lang/cem/cemcom.ansi/ch3mon.c
@@ -150,7 +150,7 @@ ch3mon(oper, expp)
 				   (arith)((*expp)->SG_LEN) :
 				   size_of_type((*expp)->ex_type,
 				       symbol2str((*expp)->ex_type->tp_fund))
-			       , INT);
+			       , ULONG);
 		expr->ex_flags |= EX_SIZEOF;
 		free_expression(*expp);
 		*expp = expr;
